% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/sample2_function.R
\name{sample2_function}
\alias{sample2_function}
\title{An example function}
\usage{
sample2_function(input = TRUE)
}
\arguments{
\item{input}{Is it TRUE/FALSE}
}
\description{
This function checks for the input argument bein passed
}
\examples{
sample2_function()
}
\keyword{sample2}

